Sarthak | Temperature is now represented as int16 and formatted as float32

+++ parser/parser_test.go
@@ -19,10 +19,10 @@ func TestParseWithTemperaturesForMinTemperature(t *testing.T) {
 	result, err := Parse(input)
 
 	assert.Nil(t, err)
-	assert.Equal(t, 6.1, result.minTemperatureOf("Canberra"))
-	assert.Equal(t, 5.9, result.minTemperatureOf("Mogadishu"))
-	assert.Equal(t, 10.2, result.minTemperatureOf("Odesa"))
-	assert.Equal(t, 9.7, result.minTemperatureOf("Tirana"))
+	assert.Equal(t, float32(6.1), result.minTemperatureOf("Canberra"))
+	assert.Equal(t, float32(5.9), result.minTemperatureOf("Mogadishu"))
+	assert.Equal(t, float32(10.2), result.minTemperatureOf("Odesa"))
+	assert.Equal(t, float32(9.7), result.minTemperatureOf("Tirana"))
 }
 
 func TestParseWithTemperaturesForMaxTemperature(t *testing.T) {
@@ -30,10 +30,10 @@ func TestParseWithTemperaturesForMaxTemperature(t *testing.T) {
 	result, err := Parse(input)
 
 	assert.Nil(t, err)
-	assert.Equal(t, 6.1, result.maxTemperatureOf("Canberra"))
-	assert.Equal(t, 6.8, result.maxTemperatureOf("Mogadishu"))
-	assert.Equal(t, 10.2, result.maxTemperatureOf("Odesa"))
-	assert.Equal(t, 15.1, result.maxTemperatureOf("Tirana"))
+	assert.Equal(t, float32(6.1), result.maxTemperatureOf("Canberra"))
+	assert.Equal(t, float32(6.8), result.maxTemperatureOf("Mogadishu"))
+	assert.Equal(t, float32(10.2), result.maxTemperatureOf("Odesa"))
+	assert.Equal(t, float32(15.1), result.maxTemperatureOf("Tirana"))
 }
 
 func TestParseWithTemperaturesForAverageTemperature(t *testing.T) {
@@ -41,8 +41,8 @@ func TestParseWithTemperaturesForAverageTemperature(t *testing.T) {
 	result, err := Parse(input)
 
 	assert.Nil(t, err)
-	assert.Equal(t, 6.1, result.averageTemperatureOf("Canberra"))
-	assert.Equal(t, 6.35, result.averageTemperatureOf("Mogadishu"))
-	assert.Equal(t, 10.2, result.averageTemperatureOf("Odesa"))
-	assert.InDelta(t, 12.4, result.averageTemperatureOf("Tirana"), 0.01)
+	assert.Equal(t, float32(6.1), result.averageTemperatureOf("Canberra"))
+	assert.Equal(t, float32(6.35), result.averageTemperatureOf("Mogadishu"))
+	assert.Equal(t, float32(10.2), result.averageTemperatureOf("Odesa"))
+	assert.InDelta(t, float32(12.4), result.averageTemperatureOf("Tirana"), 0.01)
 }
